Improve ObservedFileList::heartBeat_
Instead of just being a shared_ptr with a dummy value, we store the
this-pointer in it, which makes the closure in addWatchedDirectory
smaller since we can use the weak_ptr in favor of additionally passing
in the ObservedFileList-pointer.
As a bonus, this may make the intent of the heartBeat_ more clear.