Remove PersistentCopy().
First, we don't really need a shared_ptr from PersistentCopy(), since
we're not sharing it with anyone yet.
But then, we don't really need a pointer at all and we could just return
by value and let call-sites move the value to the heap themselves if
they need it there.
At this point, the function body is trivial, so we don't really need
the function.