Let LogFilteredDataWorkerThread::updateSearchResult() move the result
Instead of copying the result data, we exchange it for an empty array
and return the old result data. The caller can do the potentially slow
copy if necessary, or handle the new data differently, skipping a
perhaps unnecessary copy.
Currently we still do the copy, but this allows us to optimize that
step.

To reflect the new behavior, the methods dealing with the exchange have
been appropriately renamed.

There's also a small change in that the pointer-parameters are turned
into references.