[ci release] Add boolean pattern combination
This commit adds new experimental feature -- combining multiple search
patterns using boolean operators `and`, `or`, and `not` (#5, #22, #320).

This mode is enabled by separate button in the search panel. When this
mode is enabled all sub-patterns must be enclosed in `"`. `not` operator
requires its pattern to be enclosed in `()`.

Example of valid combination:
```
("pattern 1" or "pattern 2) and not("pattern 3")
```